+++ drizzle/0003_outgoing_random.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "projects" ADD COLUMN "turnstile_site_key" text;--> statement-breakpoint
+ALTER TABLE "projects" ADD COLUMN "turnstile_secret_key" text;